Book view: Only allow D-pad and circle pad in custom handler
Depending on how far you pressed the circle pad, you could activate the
normal smooth circle pad scrolling handling. Using the button list
selector completely disables that now.

May need to rework `ScrollMethod` into an enum of flags, but that seems
overkill and misleading for my one use of it in that manner...