imgui.d: Add new UI flag, `demand_focus`, and use it for the select scroll layout
This makes it so the element automatically tries to become focused when
not interacting with anything else.

This fixes an oddity where focusing on the bottom button in the book
layout will make the circle/D-pad controls do nothing. It might not be
that all scrollable boxes need to demand focus, but the two that exist
so far do.