imgui.d, imgui_render.d: Allow parents to draw offset children
Render callbacks take a `drawOffset` parameter instead of `screenPos`
and return an offset to add to all children drawn underneath them.
This is accomplished by a temporary stack that gets allocated inside
`ui.render`. I could make this dynamically growable, but I have it set
to 200 for now. This then makes that the highest depth a UI can have at
the moment.