main, imgui, imgui_render: Add bookmarking feature!
You can now select verses in the read view and bookmark them with a
button on the bottom bar. To make this work, all of the book text can
now be scrolled to the bottom screen (so you can actually select it).

Graphics are placeholder - I really need to start polishing them at this
point.