Very WIP implementation of searching
I added a "Search" button at the bottom bar in the book view that
initiates a search on a keyword through the whole Bible (naive, not
Boyer-Moore or some other fancy search algorithm) and shows a new view
of buttons that link to the results. Pressing back from one of these
linked verses will go back to the search screen first (UI view
navigation now is now kept track of with a stack).

I gave a Bible verse location a proper structure and use that in a few
places where I was manually packing into a `uint` before.