imgui: Give `Box` static member alias called `isNull`
...Since my templated functions in `util.d` that try to support any type
with an `isNull` property won't be able to see a free function defined
from another module. Namespacing a reference to the function does the
trick.

The function CANNOT be an actual member function, since `isNull` should
check if the actual struct pointer is `null`, but the compiler will
always assume member functions are not called on null struct pointers.
Trying to turn on optimizations would thus reveal pain!